requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
pytest>=7.4.3
httpx>=0.25.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4